optimizing business rules / create post route to add operation to user

+++ Challenge/app/Http/Requests/ChangeOpeningAmountRequest.php
@@ -24,7 +24,7 @@ class ChangeOpeningAmountRequest extends FormRequest
     public function rules()
     {
         return [
-            'opening_amount' => ['required', 'numeric'],
+            'opening_amount' => ['required', 'integer'],
         ];
     }
 }
